pipeline/requirements.txt
pyserial>=3.5
Pillow>=9.0
requests>=2.28
flask>=3.0
opencv-python>=4.8
google-generativeai>=0.3
